Added Display of City and Metadata
ZipContainer holds the data of the city found from the API in the Allplaces array as a state. This state is passed into ZipDIsplay. ZipDisplay was adjusted by creating an array for each city, and a seperate array for each piece of metadata. These arrays held the information, but instead formatted in JSX. Each data was then joined in another array. This array has each City and its metadata correctly formatted.